+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @core-ai/google-vertex
 
+## 0.23.0
+
+### Patch Changes
+
+- Updated dependencies [a4d08b7]
+- Updated dependencies [a4d08b7]
+    - @core-ai/google-genai@0.23.0
+    - @core-ai/core-ai@0.23.0
+
+## 0.22.0
+
+### Patch Changes
+
+- Updated dependencies [717364e]
+    - @core-ai/core-ai@0.22.0
+    - @core-ai/google-genai@0.22.0
+
+## 0.21.0
+
+### Patch Changes
+
+- Updated dependencies [37c890d]
+    - @core-ai/core-ai@0.21.0
+    - @core-ai/google-genai@0.21.0
+
 ## 0.20.0
 
 ### Patch Changes